clean up small server example
